Move event types to constants
